requirements.txt
django
djangorestframework
djangorestframework_simplejwt
stripe